feat: gpt api 활용해 QandA 객체 생성
