+++ diary/views.py
@@ -7,104 +7,19 @@ from rest_framework.viewsets import ModelViewSet
 from .models import *
 from challenge.models import *
 from rest_framework import status
-
-
-def get_diary_context(diary):
-    month = diary.month_id
-    diaries = Diary.objects.filter(month_id=month).order_by('created_date')
-
-    diary_list = list(diaries)
-    diary_index = diary_list.index(diary)
-
-    previous_diary = diary_list[diary_index - 1] if diary_index > 0 else None
-    next_diary = diary_list[diary_index + 1] if diary_index < len(diary_list) - 1 else None
-
-    previous_diary_serialized = DiarySerializer(previous_diary).data if previous_diary else None
-    next_diary_serialized = DiarySerializer(next_diary).data if next_diary else None
-    diary_serialized = DiarySerializer(diary).data
-
-    return {
-        "diary": diary_serialized,
-        "previous_diary": previous_diary_serialized,
-        "next_diary": next_diary_serialized,
-    }
-
-def complete_diary(diary_id):
-    diary = get_object_or_404(Diary, id=diary_id)
-    qandas = QandA.objects.filter(diary_id=diary.id)
-
-    if qandas.exists():
-        diary.is_complete = True
-        diary.save()
-
-        month_obj = diary.month_id
-        if month_obj:
-            month_obj.count += 1
-            month_obj.save()
-
-        daily_challenge = DailyChallenge.objects.filter(user_id=diary.user_id).first()
-        if daily_challenge:
-            daily_challenge.today_complete = True
-            daily_challenge.current_day += 1
-            daily_challenge.save()
-
-            user = User.objects.get(id=diary.user_id.id)
-            if not user.is_firstday:
-                user.is_firstday = True
-                user.save()
-
-                Badge.objects.create(
-                    title="1",
-                    type="firstday",
-                    user_id=user
-                )
-
-            if daily_challenge.current_day == daily_challenge.goal_day:
-                goal_badge_exists = Badge.objects.filter(
-                    title=str(daily_challenge.goal_day),
-                    user_id=user
-                ).exists()
-                if not goal_badge_exists:
-                    Badge.objects.create(
-                        title=str(daily_challenge.goal_day),
-                        type="goal_day",
-                        user_id=user
-                    )
-                daily_challenge.current_day = daily_challenge.goal_day
-                daily_challenge.goal_day += 7
-                daily_challenge.save()
-
-        all_daily_challenges = DailyChallenge.objects.all()
-        for challenge in all_daily_challenges:
-            if challenge.today_complete:
-                challenge.today_complete = False
-            else:
-                challenge.current_day = 0
-                challenge.goal_day = 7
-            challenge.save()
-
-        diary_context = get_diary_context(diary)
-        return {"status": "success", "data": diary_context}
-
-    return {"status": "error", "message": "Diary is not yet complete"}
+from .utils import *
 
 
 class MonthViewset(ModelViewSet):
     queryset = Month.objects.all()
     serializer_class = MonthSerializer
 
-    # month_list 불러오기 (홈 화면)
-    # @action(detail=False, methods=['get'], url_path='diary/month/<int:user_id>')
-    # def list_months(self, request, user_id=None):
-    #     queryset = self.queryset.filter(user_id=user_id)
-    #     serializer = self.get_serializer(queryset, many=True)
-    #     return Response(serializer.data)
-
 
 class DiaryViewset(ModelViewSet):
     queryset = Diary.objects.all()
     serializer_class = DiarySerializer
 
+    # complete diary
     @action(detail=True, methods=['patch'])
     def finish_diary(self, request, pk=None):
         diary = get_object_or_404(Diary, id=pk)
@@ -127,18 +42,60 @@ class DiaryViewset(ModelViewSet):
                 "message": "No QandA found. Redirecting to home screen."
             }, status=status.HTTP_200_OK)
 
-    # diary list 불러오기 (월 폴더 선택 시)
+    # 현재, 이전, 다음 다이어리 조회
     @action(detail=True, methods=['get'])
     def view_diary(self, request, pk=None):
         try:
             diary = get_object_or_404(Diary, id=pk)
             diary_context = get_diary_context(diary)
-
             return Response({"status": "success", "data": diary_context}, status=status.HTTP_200_OK)
         except Exception as e:
             return Response({"status": "error", "message": str(e)}, status=status.HTTP_500_INTERNAL_SERVER_ERROR)
+
+    # diary list 불러오기 (월 폴더 선택 시)
+    @action(detail=False, methods=['get'], url_path='list_diaries')
+    def list_diaries(self, request, month_id=None):
+        month_id = request.query_params.get('month_id')
+        if not month_id:
+            return Response({"status": "error", "message": "month_id is required"}, status=400)
+        
+        queryset = self.queryset.filter(month_id=month_id, is_complete=True)
+        serializer = self.get_serializer(queryset, many=True)
+        return Response({"status": "success", "data": serializer.data})
 
 
 class QandAViewset(ModelViewSet):
     queryset = QandA.objects.all()
     serializer_class = QandASerializer
+
+    def create(self, request, *args, **kwargs):
+        try:
+            diary_id = request.data.get('diary_id')
+            answer = request.data.get('answer')
+            if not diary_id or not answer:
+                return Response({
+                    "status": "error",
+                    "message": "diary_id and answer are required"
+                }, status=status.HTTP_400_BAD_REQUEST)
+
+            diary = get_object_or_404(Diary, id=diary_id)
+            qanda = QandA.objects.create(diary_id=diary, answer=answer)
+
+            followup_question = get_followup_question(answer)
+            qanda.question = followup_question
+            qanda.save()
+
+            # 다이어리가 완료되었는지 확인 (complete_diary 함수 호출)
+            if QandA.objects.filter(diary_id=diary.id).count() == diary.limitq_num:
+                complete_diary(diary_id)
+
+            return Response({
+                "status": "success",
+                "data": QandASerializer(qanda).data
+            }, status=status.HTTP_201_CREATED)
+
+        except Exception as e:
+            return Response({
+                "status": "error",
+                "message": str(e)
+            }, status=status.HTTP_500_INTERNAL_SERVER_ERROR)
